pyproject.toml
[tool.poetry]
name = "ELF_RSS"
version = "3.0.0-alpha.1"
description = "RSS 订阅插件，订阅源建议选择 RSSHub"
authors = ["Quan666 <i@Rori.eMail>"]
license = "GPL-3.0-only"
readme = "README.md"
packages = [
    { include = "nonebot_plugin_rss" },
]
homepage = "https://github.com/Quan666/ELF_RSS"
repository = "https://github.com/Quan666/ELF_RSS"
keywords = ["nonebot", "nonebot2", "rss" ,"elf" ,"rsshub"]

[tool.nonebot]
adapters = []
plugins = ["nonebot_plugin_rss"]

[tool.poetry.dependencies]
python = "^3.8.3"
arrow = "^1.3.0"
async-timeout = "^4.0.3"
bbcode = "^1.1.0"
cachetools = "^5.3.1"
deep-translator = "^1.11.4"
emoji = "^2.8.0"
feedparser = "^6.0.10"
ImageHash = "^4.3.1"
nonebot2 = {extras = ["fastapi", "httpx", "websockets"], version = "^2.1.2"}
nonebot-plugin-alconna = "^0.33.5"
nonebot-plugin-apscheduler = "^0.3.0"
nonebot-plugin-datastore = "^1.1.2"
nonebot-plugin-orm = {extras = ["default"], version = "^0.6.0"}
nonebot-plugin-send-anything-anywhere = "^0.4.0"
Pillow = "^10.1.0"
pydantic = "^1.10.13"
pyquery = "^2.0.0"
tenacity = "^8.2.3"
yarl = "^1.9.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.275"
black = "^23.1.0"
pre-commit = "^3.2.2"

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]

line-length = 120
target-version = "py38"

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"